requirements.txt
Flask==2.2.5
pymongo==4.5.0
PyJWT==2.7.0
flasgger==0.9.5
requests==2.31.0
bs4==0.0.1
beautifulsoup4
requests
python-dotenv
